adminsettings: Refactored workflow table classes
Added an abstract class that defines most of the column rendering.
Added workflow title display to all tables.
Fixes #20
