+++ adminlib.php
@@ -135,12 +135,16 @@ class admin_settings {
         echo $OUTPUT->heading(get_string('active_automatic_workflows_heading', 'tool_cleanupcourses'));
 
         $table = new active_automatic_workflows_table('tool_cleanupcourses_active_automatic_workflows');
+        echo $OUTPUT->box_start("cleanupcourses-enable-overflow cleanupcourses-table");
         $table->out(5000, false);
+        echo $OUTPUT->box_end();
 
         echo $OUTPUT->heading(get_string('active_manual_workflows_heading', 'tool_cleanupcourses'));
 
         $table = new active_manual_workflows_table('tool_cleanupcourses_manual_workflows');
+        echo $OUTPUT->box_start("cleanupcourses-enable-overflow cleanupcourses-table");
         $table->out(5000, false);
+        echo $OUTPUT->box_end();
 
         echo $OUTPUT->heading(get_string('workflow_definition_heading', 'tool_cleanupcourses'));
 
@@ -149,7 +153,9 @@ class admin_settings {
             get_string('add_workflow', 'tool_cleanupcourses'));
 
         $table = new workflow_definition_table('tool_cleanupcourses_workflow_definitions');
+        echo $OUTPUT->box_start("cleanupcourses-enable-overflow cleanupcourses-table");
         $table->out(5000, false);
+        echo $OUTPUT->box_end();
 
         $this->view_footer();
     }
